requirements.txt
Django>=3.2
djangorestframework~=3.12.4
django-extensions~=3.1.3
django-cors-headers
psycopg2-binary~=2.8.6
Pillow
requests
python-dateutil